fixing bug, update display, restructured directory
