+++ 15_React_Form/praktikum/src/App.tsx
@@ -2,7 +2,7 @@ import { Routes, Route } from "react-router-dom";
 import Home from "@/pages/home/page";
 import Products from "@/pages/product/page"
 import ProductDetails from "@/pages/product/details/page"
-import Header from "@/layout/header";
+import Header from "@/components/header";
 
 function App() {
   return (
@@ -10,7 +10,6 @@ function App() {
       <Header/>
       <Routes>
         <Route path="/" element={<Home />} />
-        <Route path="/home" element={<Home />} />
         <Route path="/products" element={<Products />} />
         <Route path="/products/:id" element={<ProductDetails />} />
       </Routes>
